requirementes.txt
fastapi
uvicorn
supabase
python-dotenv
jinja2
